pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fivcadvisor"
version = "0.1.0"
description = "Pattern-driven CrewAI flows and tools"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Charlie" }]
dependencies = [
  "typer>=0.12.3",
  "rich>=13.7.1",
  "PyYAML>=6.0.1",
  "python-dotenv>=1.0.1",
  "pydantic>=2.7.0",
  "crewai>=0.5.0",
  "crewai-tools[mcp]>=0.1.0",
  "fastapi>=0.104.0",
  "uvicorn[standard]>=0.24.0",
  "httpx>=0.28.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "ruff>=0.4.0,<0.6"
]

[project.scripts]
fivcadvisor = "fivcadvisor.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
